+++ src/urllib3_ext_hface/protocols/http1/__init__.py
@@ -14,6 +14,6 @@
 
 from __future__ import annotations
 
-from ._factories import HTTP1ClientFactory
+from ._h11 import HTTP1ProtocolHyperImpl
 
-__all__ = ("HTTP1ClientFactory",)
+__all__ = ("HTTP1ProtocolHyperImpl",)